Avoid "it" ARM instruction in non-Thumb mode.
This instruction only has an effect in Thumb2 mode and
some assemblers do not support this syntax.
The disadvantage is since we by default compile as ARM and
not Thumb none of the fate machines will check that the "it"
instruction encodes the same condition codes as the following
instructions.

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>

git-svn-id: http://svn.ad-group.adh/svn/native_client/native/ffmpeg/branches/adffmpeg-trunk-tracker@7147 4601b0d2-a6ba-4ff0-a3bb-59d57de6f651